Added some Functions
Added a function to retrieve information entered in the form phases.
Added a function to also retrieve images as well.
Added a function to check if all form fields are filled before proceeding to next phase. This function also displays an error message to tell the user to fill in all fields. Added some styling for the error message as well